remove the loopfor server.run and clean up
