+++ Cargo.toml
@@ -49,6 +49,7 @@ default-features = false
 features = ["matrix_graph"]
 
 # =========== Package configuration ===========
+
 [package]
 name = "calyx"
 build = false
